# BitFilter made faster