--- a/src/Paprika.Benchmarks/PageExtensionsBenchmarks.cs
+++ b/src/Paprika.Benchmarks/PageExtensionsBenchmarks.cs
@@ -1,0 +1,23 @@
+﻿using BenchmarkDotNet.Attributes;
+using Paprika.Data;
+using Paprika.Store;
+
+namespace Paprika.Benchmarks;
+
+public class PageExtensionsBenchmarks
+{
+    private readonly Page _a;
+    private readonly Page _b;
+
+    public unsafe PageExtensionsBenchmarks()
+    {
+        _a = new Page((byte*)Allocator.AllocAlignedPage());
+        _b = new Page((byte*)Allocator.AllocAlignedPage());
+    }
+
+    [Benchmark]
+    public void OrWith()
+    {
+        _a.OrWith(_b);
+    }
+}